feat/fix: created CSSAPageCards, which don't get externally linked
