+++ src/app/resources/page.tsx
@@ -8,7 +8,7 @@ import {
     CarouselNext,
     CarouselPrevious
 } from "@/components/ui/carousel";
-import { CSSALinks, IPlaylist, MeetingArchivesID, PlaylistCollections, ResourceLinks } from "@/data/resources";
+import { CSSALinks, CSSAPageLinks, IPlaylist, MeetingArchivesID, PlaylistCollections, ResourceLinks } from "@/data/resources";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from "@/components/ui/accordion";
 
@@ -17,7 +17,6 @@ import { GetPlaylistData, IVideoData } from "@/api/youtube";
 import React from "react";
 import { FaExternalLinkAlt } from "react-icons/fa";
 import Link from "next/link";
-import { getAllGuides } from "@/lib/mdx";
 
 function MakePlaylistCards(videos: IVideoData[]) {
     return videos?.map((video, index) => {
@@ -49,10 +48,6 @@ export default async function Resources() {
             playlist.videos = videos;
         }
     }
-    
-    // Get guides count
-    const guides = await getAllGuides();
-    const guidesCount = guides.length;
 
     const resourceCards = ResourceLinks.map((link, index) => {
         return (
@@ -92,35 +87,28 @@ export default async function Resources() {
         );
     });
 
-    // Add guides card
-    const guidesCard = (
-        <CarouselItem className="md:basis-1/2 lg:basis-[30%]">
-            <div className="px-2 w-full h-full">
-                <Link href="/resources/guides">
-                    <Card className="h-full border-primary">
-                        <CardHeader>
-                            <CardTitle className="text-lg flex flex-row gap-2">
-                                Student Guides
-                            </CardTitle>
-                            <CardDescription>
-                                Collection of {guidesCount} comprehensive guides for computer science students
-                            </CardDescription>
-                        </CardHeader>
-                        <CardContent>
-                            <div className="flex justify-end">
-                                <span className="text-sm text-primary hover:underline">
-                                    Browse guides →
-                                </span>
-                            </div>
-                        </CardContent>
-                    </Card>
-                </Link>
-            </div>
-        </CarouselItem>
-    );
+    // Page cards are not externally linked
+    const cssaPageCards = CSSAPageLinks.map((link, index) => {
+        return (
+            <CarouselItem key={index} className="md:basis-1/2 lg:basis-[30%]" >
+                <div className="px-2 w-full h-full">
+                    <a href={link.href} rel="noreferrer">
+                        <Card className="h-full">
+                            <CardHeader>
+                                <CardTitle className="text-lg flex flex-row gap-2">
+                                    {link.title} <FaExternalLinkAlt className="my-auto" />
+                                </CardTitle>
+                                <CardDescription>{link.description}</CardDescription>
+                            </CardHeader>
+                        </Card>
+                    </a>
+                </div>
+            </CarouselItem>
+        );
+    });
 
-    // Add guides card to CSSA resources
-    const allCssaCards = [guidesCard, ...cssaCards];
+    // combine page cards and regular cards
+    const allCssaCards = cssaPageCards.concat(cssaCards);
 
     const collectionTabs = PlaylistCollections.map((collection, index) => {
         return (
