--- src/data/guides.ts
@@ -1,18 +1,0 @@
-interface IGuideList {
-    slug: string;
-    repoURL: string;
-}
-
-const GuideList: IGuideList[] = [
-    {
-        slug: 'first-year-guide',
-        repoURL: 'https://github.com/umanitoba-cssa/first-year-guide',
-    },
-    {
-        slug: 'second-year-guide',
-        repoURL: 'https://github.com/umanitoba-cssa/second-year-guide'
-    },
-]
-
-export { GuideList };
-export default GuideList;
